move all unsafe publication tests into single file
