+++ testsuite/src/commonMain/kotlin/org/jetbrains/litmuskt/tests/UnsafePublication.kt
@@ -4,6 +4,7 @@ import org.jetbrains.litmuskt.LitmusIOutcome
 import org.jetbrains.litmuskt.LitmusTestContainer
 import org.jetbrains.litmuskt.accept
 import org.jetbrains.litmuskt.litmusTest
+import kotlin.concurrent.Volatile
 
 data class IntHolder(val x: Int)
 
@@ -14,6 +15,24 @@ object UnsafePublication {
 
     val Plain = litmusTest({
         object : LitmusIOutcome() {
+            var h: IntHolder? = null
+        }
+    }) {
+        thread {
+            h = IntHolder(0)
+        }
+        thread {
+            r1 = h?.x ?: -1
+        }
+        spec {
+            accept(0)
+            accept(-1)
+        }
+    }
+
+    val Volatile = litmusTest({
+        object : LitmusIOutcome() {
+            @Volatile
             var h: IntHolder? = null
         }
     }) {
@@ -42,6 +61,76 @@ object UnsafePublication {
         }
         spec {
             accept(1)
+            accept(-1)
+        }
+    }
+
+    val Array = litmusTest({
+        object : LitmusIOutcome() {
+            var arr: Array<Int>? = null
+        }
+    }) {
+        thread {
+            arr = Array(10) { 0 }
+        }
+        thread {
+            r1 = arr?.get(0) ?: -1
+        }
+        spec {
+            accept(0)
+            accept(-1)
+        }
+    }
+
+    private class UPUBRefInner(val x: Int)
+    private class UPUBRefHolder(val ref: UPUBRefInner)
+
+    val Reference = litmusTest({
+        object : LitmusIOutcome() {
+            var h: UPUBRefHolder? = null
+        }
+    }) {
+        thread {
+            val ref = UPUBRefInner(1)
+            h = UPUBRefHolder(ref)
+        }
+        thread {
+            val t = h
+            r1 = t?.ref?.x ?: -1
+        }
+        spec {
+            accept(1)
+            accept(-1)
+        }
+    }
+
+    private class UPUBIntHolderInnerLeaking {
+        var ih: InnerHolder? = null
+
+        inner class InnerHolder {
+            val x: Int
+
+            init {
+                x = 1
+                ih = this
+            }
+        }
+    }
+
+    val ctorLeaking = litmusTest({
+        object : LitmusIOutcome() {
+            var h = UPUBIntHolderInnerLeaking()
+        }
+    }) {
+        thread {
+            h.InnerHolder()
+        }
+        thread {
+            r1 = h.ih?.x ?: -1
+        }
+        spec {
+            accept(1)
+            accept(0)
             accept(-1)
         }
     }
